% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{fast_build_index}
\alias{fast_build_index}
\title{Build a Reusable Subject Index}
\usage{
fast_build_index(subject)
}
\arguments{
\item{subject}{An \code{IRanges} or \code{GRanges} object.}
}
\value{
A \code{fast_ranges_index} object.
}
\description{
Build a sorted subject index that can be reused across repeated overlap
queries.
}
\examples{
s <- IRanges::IRanges(start = c(3, 9, 18), width = c(4, 6, 5))
idx <- fast_build_index(s)
idx
}
