% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_ops.R
\name{fast_disjoin}
\alias{fast_disjoin}
\title{Disjoin Ranges}
\usage{
fast_disjoin(x, ignore_strand = FALSE, with_revmap = FALSE)
}
\arguments{
\item{x}{An \code{IRanges} or \code{GRanges} object.}

\item{ignore_strand}{Logical scalar. Ignored for non-genomic ranges.}

\item{with_revmap}{Logical scalar. If \code{TRUE}, include reverse mapping.}
}
\value{
An object of the same range class as \code{x}.
}
\description{
Return non-overlapping segments induced by input ranges.
}
\examples{
x <- IRanges::IRanges(start = c(1, 4, 10), end = c(5, 8, 12))
fast_disjoin(x)
}
