% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{fast_iter_reset}
\alias{fast_iter_reset}
\title{Reset an Overlap Iterator}
\usage{
fast_iter_reset(iter)
}
\arguments{
\item{iter}{A \code{fast_ranges_iter} object.}
}
\value{
Invisibly returns \code{iter}.
}
\description{
Reset iterator position to the first query chunk.
}
\examples{
q <- IRanges::IRanges(start = c(1, 10, 20), width = c(5, 4, 3))
s <- IRanges::IRanges(start = c(3, 9, 18), width = c(4, 6, 5))
it <- fast_find_overlaps_iter(q, s, chunk_size = 2L)
fast_iter_reset(it)
}
