% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_ops.R
\name{fast_range_union}
\alias{fast_range_union}
\title{Union of Two Range Sets}
\usage{
fast_range_union(x, y, ignore_strand = FALSE)
}
\arguments{
\item{x, y}{\code{IRanges} or \code{GRanges} objects of compatible class.}

\item{ignore_strand}{Logical scalar. Ignored for non-genomic ranges.}
}
\value{
An object of the same range class as \code{x} and \code{y}.
}
\description{
Compute range-wise union.
}
\details{
\code{fast_range_union()} returns the combined interval coverage of \code{x} and \code{y}.
}
\examples{
x <- IRanges::IRanges(start = c(1, 10), end = c(5, 12))
y <- IRanges::IRanges(start = c(3, 20), end = c(8, 21))
fast_range_union(x, y)
}
