% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_ops.R
\name{fast_reduce}
\alias{fast_reduce}
\title{Reduce Overlapping Ranges}
\usage{
fast_reduce(x, ignore_strand = FALSE, min_gap_width = 1L, with_revmap = FALSE)
}
\arguments{
\item{x}{An \code{IRanges} or \code{GRanges} object.}

\item{ignore_strand}{Logical scalar. Ignored for non-genomic ranges.}

\item{min_gap_width}{Integer scalar minimum gap width for separating ranges.}

\item{with_revmap}{Logical scalar. If \code{TRUE}, include reverse mapping.}
}
\value{
An object of the same range class as \code{x}.
}
\description{
Merge overlapping or adjacent ranges.
}
\examples{
x <- IRanges::IRanges(start = c(1, 4, 10), end = c(5, 8, 12))
fast_reduce(x)
}
